Refactor domain analysis sections into dedicated components
