+++ cloud/apps/web/src/components/domains/domainAnalysisColors.ts
@@ -1,0 +1,33 @@
+export function getHeatmapColor(value: number): string {
+  const clamped = Math.max(-1, Math.min(1, value));
+  const stops = [
+    { at: -1, rgb: [153, 27, 27] },
+    { at: -0.6, rgb: [239, 68, 68] },
+    { at: -0.2, rgb: [254, 202, 202] },
+    { at: 0, rgb: [250, 204, 21] },
+    { at: 0.44, rgb: [134, 239, 172] },
+    { at: 0.76, rgb: [22, 163, 74] },
+    { at: 1, rgb: [21, 128, 61] },
+  ] as const;
+
+  const rightIndex = stops.findIndex((stop) => clamped <= stop.at);
+  if (rightIndex <= 0) {
+    const [r, g, b] = stops[0].rgb;
+    return `rgba(${r}, ${g}, ${b}, 0.35)`;
+  }
+
+  const left = stops[rightIndex - 1]!;
+  const right = stops[rightIndex]!;
+  const localT = (clamped - left.at) / (right.at - left.at);
+  const r = Math.round(left.rgb[0] + (right.rgb[0] - left.rgb[0]) * localT);
+  const g = Math.round(left.rgb[1] + (right.rgb[1] - left.rgb[1]) * localT);
+  const b = Math.round(left.rgb[2] + (right.rgb[2] - left.rgb[2]) * localT);
+  return `rgba(${r}, ${g}, ${b}, 0.35)`;
+}
+
+export function getPriorityColor(value: number, min: number, max: number): string {
+  if (max <= min) return getHeatmapColor(0);
+  const normalized = (value - min) / (max - min);
+  const heatmapScale = normalized * 2 - 1;
+  return getHeatmapColor(heatmapScale);
+}
